Move the base query object
